+++ xonsh/parser/completion_context.py
